% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{delta_heatmap}
\alias{delta_heatmap}
\title{Interactive Heatmaps for Delta-Scale Pairwise Correlations}
\usage{
delta_heatmap(x, timediff = "2-1", interactive = TRUE, grayscale = FALSE)
}
\arguments{
\item{x}{Input covariate array, with n rows, p columns, and t slices}

\item{timediff}{Pair of time points to use in heatmap. Should be in format \code{'t2-t1'}}

\item{interactive}{If \code{TRUE}, a shiny app will open in browser that will display an interactive version of the heatmap where sub-heatmaps can be selected for display}

\item{grayscale}{If \code{TRUE} the viridis colors will be desaturated to grayscale}
}
\value{
A heatmap, possibly interactive in a shiny app
}
\description{
Interactive Heatmaps for Delta-Scale Pairwise Correlations
}
\examples{
\dontrun{
delta_heatmap(Xarray)
delta_heatmap(Xarray, timediff = "4-2", interactive = F, grayscale = T)
}
}
