% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{delta_scatter}
\alias{delta_scatter}
\title{2D and 3D Scatter Plots for Delta-Scale Pairwise Correlations}
\usage{
delta_scatter(
  x,
  timediff1 = "2-1",
  timediff2 = "3-2",
  timediff3 = NULL,
  fisherz = TRUE,
  interactive = TRUE,
  digits = 3
)
}
\arguments{
\item{x}{Input covariate array, with n rows, p columns, and t slices}

\item{timediff1}{First pair of time points for the x-axis of the scatter plot. Should be in format \code{'t2-t1'}}

\item{timediff2}{Second pair of time points for the y-axis of the scatter plot. Should be in format \code{'t2-t1'}}

\item{timediff3}{Optional third pair of time points for the z-axis of the scatter plot. Should be in format \code{'t2-t1'}}

\item{fisherz}{If \code{TRUE}, the Fisher z-transformation (atanh) will be applied to the correlations}

\item{interactive}{If \code{TRUE} and timediff3 is \code{NULL}, an interactive \code{plotly} plot with hover text containing correlations and variable names will be generated. If \code{FALSE} and timediff3 is \code{NULL}, a regular \code{ggplot2} scatterplot will be shown. If timediff3 is provided, this parameter will only determine whether hovertext is shown}

\item{digits}{Number of digits for rounding in the hovertext}
}
\value{
Either a 2D or 3D scatterplot
}
\description{
2D and 3D Scatter Plots for Delta-Scale Pairwise Correlations
}
\examples{
\dontrun{
delta_scatter(Xarray)
delta_scatter(Xarray, timediff3 = "4-3")
delta_scatter(Xarray, timediff1 = "3-1", timediff2 = "5-3", timediff3 = "7-5")
}

\insertRef{ggplot2}{prolong}
\insertRef{plotly}{prolong}
}
