% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prolong.R
\name{prolong}
\alias{prolong}
\title{Fit prolong Model}
\usage{
prolong(
  Y,
  X,
  lambda1 = NULL,
  lambda2 = NULL,
  lambdar = NULL,
  groups = TRUE,
  foldids = NULL,
  optimvals = c(1, 0.01)
)
}
\arguments{
\item{Y}{Input response matrix, with n rows and t columns}

\item{X}{Input covariate array, with n rows, p columns, and t slices}

\item{lambda1}{Lasso/group lasso parameter, if left \code{NULL} this parameter
will be chosen via a cross-validation that keeps each subject's observations
across time points together. It is recommended to save the lambda2 and
lambdar values from the first run of prolong for future runs, since the
optimization step takes the longest}

\item{lambda2}{Laplacian penalty parameter, if left \code{NULL} will be chosen along
with lambdar via MLE, see supplementary material in
\insertCite{spatial-connectivity}{prolong}}

\item{lambdar}{Nuisance parameter, if left \code{NULL} will be chosen along with
lambdar via MLE. Added to the the diagonal elements of the laplacian matrix
to get the invertibility required for the MLE}

\item{groups}{Optional pre-specified groups. If \code{NULL} or \code{FALSE}, lasso will be
used. If left as \code{TRUE}, then there will be p groups each containing the
observations across time points}

\item{foldids}{Optional pre-specified foldids for the cv. Should be of length
n. If left \code{NULL}, subjects will be automatically split into 5 folds}

\item{optimvals}{Initial values of lambda2 and lambdar to be used by
\code{optim()} for the MLE of lambda2 and lambdar}
}
\value{
A named numeric vector of coefficients from the lasso or group lasso
}
\description{
\code{prolong()} is designed to take an \eqn{n \times t} outcome matrix and an
\eqn{n \times p \times t} array of covariates and automatically run the
entire processing and model fitting process, returning a list of coefficients
with corresponding variable names.
}
\details{
First, the data is reshaped into the first-differenced vectorized
\eqn{n(t-1)} length Y and first-differenced matricized \eqn{n(t-1) \times
pt(t-1)/2} X with a corresponding dependence matrix.
Next, hyperparameters for the graph laplacian-based network
penalty are found via MLE and the parameter for lasso/group lasso is found
via a careful implementation of cross-validation.
Lastly, a group lasso + laplacian or lasso + laplacian model is implemented,
and its bias-adjusted coefficients are returned.
}
\examples{
\dontrun{
coefs <- prolong(Ymatrix, Xarray)
coefs

coefs <- prolong(Ymatrix, Xarray, lambda2 = .001, lambdar = 10, groups = )
coefs
}

}
\references{
\insertAllCited{}
\insertRef{gglasso}{prolong}
\insertRef{group-lasso}{prolong}
}
