% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_metabs}
\alias{sim_metabs}
\title{Simulated Metabolite Data}
\format{
\subsection{\code{sim_metabs}}{

A 3-D numeric array with n = 15 rows, p = 100 columns, and t = 4 slices
}
}
\usage{
data(sim_metabs)
}
\description{
A 15x100x4 array containing 20 simulated target metabolites and 80 simulated
noise metabolites across 15 observations and 4 time points. Intended for use
in \code{prolong} package plotting functions or as predictors for \code{sim_outcome} in
the \code{prolong()} function.
}
\keyword{datasets}
