% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_outcome}
\alias{sim_outcome}
\title{Simulated Clinical Outcome Data}
\format{
\subsection{\code{sim_outcome}}{

A numeric matrix with n = 15 rows and t = 4 columns
}
}
\usage{
data(sim_metabs)
}
\description{
A 15x4 matrix containing 20 simulated clinical outcome values. Intended for
use as the response variable with predictors from \code{sim_metabs} in
the \code{prolong()} function.
}
\keyword{datasets}
