% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{delta_heatmap}
\alias{delta_heatmap}
\title{Interactive Heatmaps for Delta-Scale Pairwise Correlations}
\usage{
delta_heatmap(
  x,
  timediff = "2-1",
  object = NULL,
  selected = NULL,
  interactive = TRUE,
  grayscale = FALSE
)
}
\arguments{
\item{x}{Input covariate array, with n rows, p columns, and t slices}

\item{timediff}{Pair of time points to use in heatmap. Should be in format \code{'t2-t1'}}

\item{object}{A \code{prolong} model object. Can be left \code{NULL} if \code{selected} is provided or if all variables are to be used}

\item{selected}{A character list of variable names or a numeric index of variables of interest to be included in the heatmap Can be left \code{NULL} if object is provided or if all variables are to be used}

\item{interactive}{If \code{TRUE}, a shiny app will open in browser that will display an interactive version of the heatmap where sub-heatmaps can be selected for display}

\item{grayscale}{If \code{TRUE} the viridis colors will be desaturated to grayscale}
}
\value{
A heatmap, possibly interactive in a shiny app
}
\description{
\code{delta_heatmap()} is designed to take an \eqn{n \times p \times t} array
and display a heatmap to visualize the observed delta-scalecorrelation
matrix. This delta-scale correlation matrix is used in the \code{prolong} model as
the adjacency matrix for the graph whose laplacian is used in the (group)
lasso + laplacian penalty used.
}
\examples{
\dontrun{
delta_heatmap(Xarray)
delta_heatmap(Xarray, timediff = "4-2", interactive = F, grayscale = T)
}

}
\references{
\insertRef{CH2}{prolong}
\insertRef{CH3}{prolong}
}
