% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prolong.R
\name{prolong}
\alias{prolong}
\title{Title}
\usage{
prolong(
  Y,
  X,
  lambda1 = NULL,
  lambda2 = NULL,
  lambdar = NULL,
  groups = TRUE,
  foldids = NULL
)
}
\arguments{
\item{Y}{Input response matrix, with n rows and t columns}

\item{X}{Input covariate array, with n rows, p columns, and t slices}

\item{lambda1}{Lasso/group lasso parameter, if left NULL will be chosen via a
cross-validation that keeps each subject's observations across time points
together. It is recommended to save the lambda2 and lambdar values from the
first run of prolong for future runs, since the optimization step takes the
longest}

\item{lambda2}{Laplacian penalty parameter, if left NULL will be chosen along
with lambdar via MLE}

\item{lambdar}{Nuisance parameter, if left NULL will be chosen along with
lambdar via MLE. Added to the the diagonal elements of the laplacian matrix
 to get the invertibility required for the MLE}

\item{groups}{Optional pre-specified groups. If NULL or FALSE, lasso will be
used. If left as TRUE, then there will be p groups each containing the
 observations across time points}

\item{foldids}{Optional pre-specified foldids for the cv. Should be of length
n. If left NULL, subjects will be automatically split into 5 folds}
}
\value{
A named numeric vector of coefficients from the lasso or group lasso
}
\description{
Title
}
\examples{
\dontrun{
coefs <- prolong(Ymatrix, Xarray)
coefs

coefs <- prolong(Ymatrix, Xarray, lambda2 = .001, lambdar = 10, groups = )
coefs
}
}
